--- FF1Lib/Treasures.txt
@@ -1,256 +1,0 @@
-"Unused"
-"Coneria 1"
-"Coneria 2"
-"Coneria 3"
-"Coneria 4"
-"Coneria 5"
-"Coneria 6"
-"Temple of Fiends 1"
-"Temple of Fiends 2"
-"Temple of Fiends 3"
-"Temple of Fiends 4"
-"Temple of Fiends 5"
-"Temple of Fiends 6"
-"Elfland 1"
-"Elfland 2"
-"Elfland 3"
-"Elfland 4"
-"NorthWest Castle 1"
-"NorthWest Castle 2"
-"NorthWest Castle 3"
-"Marsh Cave 1"
-"Marsh Cave 2"
-"Marsh Cave 3"
-"Marsh Cave 4"
-"Marsh Cave 5"
-"Marsh Cave 6"
-"Marsh Cave 7"
-"Marsh Cave 8"
-"Marsh Cave 9"
-"Marsh Cave 10"
-"Marsh Cave 11"
-"Marsh Cave 12"
-"Marsh Cave 13"
-"Dwarf Cave 1"
-"Dwarf Cave 2"
-"Dwarf Cave 3"
-"Dwarf Cave 4"
-"Dwarf Cave 5"
-"Dwarf Cave 6"
-"Dwarf Cave 7"
-"Dwarf Cave 8"
-"Dwarf Cave 9"
-"Dwarf Cave 10"
-"Matoya's Cave 1"
-"Matoya's Cave 2"
-"Matoya's Cave 3"
-"Earth Cave 1"
-"Earth Cave 2"
-"Earth Cave 3"
-"Earth Cave 4"
-"Earth Cave 5"
-"Earth Cave 6"
-"Earth Cave 7"
-"Earth Cave 8"
-"Earth Cave 9"
-"Earth Cave 10"
-"Earth Cave 11"
-"Earth Cave 12"
-"Earth Cave 13"
-"Earth Cave 14"
-"Earth Cave 15"
-"Earth Cave 16"
-"Earth Cave 17"
-"Earth Cave 18"
-"Earth Cave 19"
-"Earth Cave 20"
-"Earth Cave 21"
-"Earth Cave 22"
-"Earth Cave 23"
-"Earth Cave 24"
-"Titan's Tunnel 1"
-"Titan's Tunnel 2"
-"Titan's Tunnel 3"
-"Titan's Tunnel 4"
-"Gurgu Volcano 1"
-"Gurgu Volcano 2"
-"Gurgu Volcano 3"
-"Gurgu Volcano 4"
-"Gurgu Volcano 5"
-"Gurgu Volcano 6"
-"Gurgu Volcano 7"
-"Gurgu Volcano 8"
-"Gurgu Volcano 9"
-"Gurgu Volcano 10"
-"Gurgu Volcano 11"
-"Gurgu Volcano 12"
-"Gurgu Volcano 13"
-"Gurgu Volcano 14"
-"Gurgu Volcano 15"
-"Gurgu Volcano 16"
-"Gurgu Volcano 17"
-"Gurgu Volcano 18"
-"Gurgu Volcano 19"
-"Gurgu Volcano 20"
-"Gurgu Volcano 21"
-"Gurgu Volcano 22"
-"Gurgu Volcano 23"
-"Gurgu Volcano 24"
-"Gurgu Volcano 25"
-"Gurgu Volcano 26"
-"Gurgu Volcano 27"
-"Gurgu Volcano 28"
-"Gurgu Volcano 29"
-"Gurgu Volcano 30"
-"Gurgu Volcano 31"
-"Gurgu Volcano 32"
-"Gurgu Volcano 33"
-"Ice Cave 1"
-"Ice Cave 2"
-"Ice Cave 3"
-"Ice Cave 4"
-"Ice Cave 5"
-"Ice Cave 6"
-"Ice Cave 7"
-"Ice Cave 8"
-"Ice Cave 9"
-"Ice Cave 10"
-"Ice Cave 11"
-"Ice Cave 12"
-"Ice Cave 13"
-"Ice Cave 14"
-"Ice Cave 15"
-"Ice Cave 16"
-"Castle of Ordeal 1"
-"Castle of Ordeal 2"
-"Castle of Ordeal 3"
-"Castle of Ordeal 4"
-"Castle of Ordeal 5"
-"Castle of Ordeal 6"
-"Castle of Ordeal 7"
-"Castle of Ordeal 8"
-"Castle of Ordeal 9"
-"Cardia 1"
-"Cardia 2"
-"Cardia 3"
-"Cardia 4"
-"Cardia 5"
-"Cardia 6"
-"Cardia 7"
-"Cardia 8"
-"Cardia 9"
-"Cardia 10"
-"Cardia 11"
-"Cardia 12"
-"Cardia 13"
-"Not Used 1"
-"Not Used 2"
-"Not Used 3"
-"Not Used 4"
-"Sea Shrine 1"
-"Sea Shrine 2"
-"Sea Shrine 3"
-"Sea Shrine 4"
-"Sea Shrine 5"
-"Sea Shrine 6"
-"Sea Shrine 7"
-"Sea Shrine 8"
-"Sea Shrine 9"
-"Sea Shrine 10"
-"Sea Shrine 11"
-"Sea Shrine 12"
-"Sea Shrine 13"
-"Sea Shrine 14"
-"Sea Shrine 15"
-"Sea Shrine 16"
-"Sea Shrine 17"
-"Sea Shrine 18"
-"Sea Shrine 19"
-"Sea Shrine 20"
-"Sea Shrine 21"
-"Sea Shrine 22"
-"Sea Shrine 23"
-"Sea Shrine 24"
-"Sea Shrine 25"
-"Sea Shrine 26"
-"Sea Shrine 27"
-"Sea Shrine 28"
-"Sea Shrine 29"
-"Sea Shrine 30"
-"Sea Shrine 31"
-"Sea Shrine 32"
-"Waterfall 1"
-"Waterfall 2"
-"Waterfall 3"
-"Waterfall 4"
-"Waterfall 5"
-"Waterfall 6"
-"Not Used 5"
-"Not Used 6"
-"Not Used 7"
-"Not Used 8"
-"Not Used 9"
-"Not Used 10"
-"Not Used 11"
-"Not Used 12"
-"Not Used 13"
-"Mirage Tower 1"
-"Mirage Tower 2"
-"Mirage Tower 3"
-"Mirage Tower 4"
-"Mirage Tower 5"
-"Mirage Tower 6"
-"Mirage Tower 7"
-"Mirage Tower 8"
-"Mirage Tower 9"
-"Mirage Tower 10"
-"Mirage Tower 11"
-"Mirage Tower 12"
-"Mirage Tower 13"
-"Mirage Tower 14"
-"Mirage Tower 15"
-"Mirage Tower 16"
-"Mirage Tower 17"
-"Mirage Tower 18"
-"Sky Palace 1"
-"Sky Palace 2"
-"Sky Palace 3"
-"Sky Palace 4"
-"Sky Palace 5"
-"Sky Palace 6"
-"Sky Palace 7"
-"Sky Palace 8"
-"Sky Palace 9"
-"Sky Palace 10"
-"Sky Palace 11"
-"Sky Palace 12"
-"Sky Palace 13"
-"Sky Palace 14"
-"Sky Palace 15"
-"Sky Palace 16"
-"Sky Palace 17"
-"Sky Palace 18"
-"Sky Palace 19"
-"Sky Palace 20"
-"Sky Palace 21"
-"Sky Palace 22"
-"Sky Palace 23"
-"Sky Palace 24"
-"Sky Palace 25"
-"Sky Palace 26"
-"Sky Palace 27"
-"Sky Palace 28"
-"Sky Palace 29"
-"Sky Palace 30"
-"Sky Palace 31"
-"Sky Palace 32"
-"Sky Palace 33"
-"Sky Palace 34"
-"ToF Revisited 1"
-"ToF Revisited 2"
-"ToF Revisited 3"
-"ToF Revisited 4"
-"ToF Revisited 5"
-"ToF Revisited 6"
-"ToF Revisited 7"
-"Unused"
